% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_FOOOF.R
\name{apped_EEG}
\alias{apped_EEG}
\title{Combine EEG model data}
\usage{
apped_EEG(file_path, output_name, conds_pattern = ".._[0-9][0-9]")
}
\arguments{
\item{file_path}{Path to directory of .csv files to rbind. Nothing else should
be in this directory.}

\item{output_name}{Name/name and path to output file.}

\item{conds_pattern}{This function assumes a common naming convention among
.csv files. For example, if all .csv files for the Eyes Closed condition at
32Hz contain "EC_32" in the file name, then the conds_pattern argument will
take a regex is an input to parse the .csv file name and create a column for
the output .csv, called \code{condition}, which extracts that regex. Defaults to
"..\emph{\link{0-9}" which parses "XX}##".}
}
\description{
This function is intended to be used after \code{FOOOF_summaries()} and is fairly
inflexible The output from \code{FOOOF_summaries()} is a .csv file written to some data
directory. If the \code{FOOOF_summaries()} function was implemented a number times,
numerous .csv files will have been generated for each EEG condition, for
example. This function appends (rbinds) those model data .csv files in a
\emph{single} directory. If the files are spread across directories, they will all
need to be moved to one directory, and they should be the \emph{only} files in that
directory. All data will need to have the same variables (columns) for this
function to work.
}
\keyword{EEG}
