% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOOOF_report.R
\name{FOOOF_report}
\alias{FOOOF_report}
\title{Run model fit, and display a report, which includes a plot, and printed results.}
\usage{
FOOOF_report(
  fooof_obj,
  freqs = NULL,
  power_spectrum = NULL,
  freq_range = NULL,
  plt_log = FALSE
)
}
\arguments{
\item{freqs}{freqs}

\item{power_spectrum}{power_spectrum}

\item{freq_range}{freq_range}

\item{plt_log}{plt_log}
}
\description{
\subsection{Parameters}{
}

\subsection{freqs : 1d array, optional Frequency values for the power spectrum.
power_spectrum : 1d array, optional Power values, which must be input in linear space.
freq_range : list of \link{float, float}, optional Desired frequency range to fit the model to. If not provided, fits across the entire given range.
plt_log : bool, optional, default: FALSE Whether or not to plot the frequency axis in log space. Notes}{

Data is optional, if data has already been added to the object.
}
}
\section{freqs }{
 1d array, optional:
Frequency values for the power spectrum. power_spectrum : 1d array, optional Power values, which must be input in linear space. freq_range : list of \link{float, float}, optional Desired frequency range to fit the model to. If not provided, fits across the entire given range. plt_log : bool, optional, default: FALSE Whether or not to plot the frequency axis in log space.
}

\section{power_spectrum }{
 1d array, optional:
Power values, which must be input in linear space. freq_range : list of \link{float, float}, optional Desired frequency range to fit the model to. If not provided, fits across the entire given range. plt_log : bool, optional, default: FALSE Whether or not to plot the frequency axis in log space.
}

\section{freq_range }{
 list of \link{float, float}, optional:
Desired frequency range to fit the model to. If not provided, fits across the entire given range. plt_log : bool, optional, default: FALSE Whether or not to plot the frequency axis in log space.
}

\section{plt_log }{
 bool, optional, default: False:
Whether or not to plot the frequency axis in log space.
}

