% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOOOF_summaries.R
\name{FOOOF_summaries}
\alias{FOOOF_summaries}
\title{Summary function implementing FOOOF module in Python}
\usage{
FOOOF_summaries(
  data_list,
  append = NULL,
  img_path = NULL,
  data_path = NULL,
  freq_range = c(1L, 50L)
)
}
\arguments{
\item{data_list}{data_list}

\item{append}{append}

\item{img_path}{img_path}

\item{data_path}{data_path}

\item{freq_range}{freq_range}
}
\description{
Summary function implementing FOOOF module in Python
}
