% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EEG.R
\name{read_EEG}
\alias{read_EEG}
\title{Read EEG excel files}
\usage{
read_EEG(file, skip = 2:length(sheets), verbose = TRUE)
}
\arguments{
\item{file}{Path to file to be read in.}

\item{skip}{Numeric or character vector defining sheets to be skipped when reading.
Defaults to 2:length(sheets).}

\item{verbose}{Provides a summary of what was read in. Defaults to True.}
}
\description{
This function reads in EEG excel files where each sheet is a participant,
each electrode is a row and each is the power expressed in Hz.
}
\keyword{EEG}
